Create code.html
This is some jquery that can let you edit html-text straight into your html document. edit-in-place i think the people call this kind of functionality. 

Kinda mimmics the functionality in jEditable or something. But, i knew i could do it in fewer lines and more specific to my needs. So... yeah...